+++ src/main/java/org/olat/modules/portfolio/manager/PortfolioServiceImpl.java
@@ -349,9 +349,7 @@ public class PortfolioServiceImpl implements PortfolioService {
 		File reFolder = frm.getFileResourceRoot(resource);
 		File configXml = new File(reFolder, PACKAGE_CONFIG_FILE_NAME);
 		if(options == null) {
-			if(configXml.exists()) {
-				configXml.delete();
-			}
+			FileUtils.deleteFile(configXml);
 		} else {
 			try (OutputStream out = new FileOutputStream(configXml)) {
 				configXstream.toXML(options, out);
@@ -812,22 +810,39 @@ public class PortfolioServiceImpl implements PortfolioService {
 	}
 	
 	@Override
-	public void removeAccessRights(Binder binder, Identity identity) {
+	public void removeAccessRights(Binder binder, Identity identity, PortfolioRoles... roles) {
+		Set<PortfolioRoles> roleSet = new HashSet<>();
+		if(roles != null && roles.length > 0) {
+			for(PortfolioRoles role:roles) {
+				if(role != null) {
+					roleSet.add(role);
+				}
+			}
+		}
+		
+		if(roleSet.isEmpty()) {
+			log.warn("Want to remove rights without specifying the roles.");
+			return;
+		}
+
 		List<AccessRights> rights = getAccessRights(binder, identity);
 		for(AccessRights right:rights) {
-			Group baseGroup;
-			if(right.getType() == PortfolioElementType.binder) {
-				baseGroup = binderDao.loadByKey(right.getBinderKey()).getBaseGroup();
-			} else if(right.getType() == PortfolioElementType.section) {
-				baseGroup = binderDao.loadSectionByKey(right.getSectionKey()).getBaseGroup();
-			} else if(right.getType() == PortfolioElementType.page) {
-				baseGroup = pageDao.loadByKey(right.getPageKey()).getBaseGroup();
-			} else {
-				continue;
-			}
-			
-			if(groupDao.hasRole(baseGroup, identity, right.getRole().name())) {
-				groupDao.removeMembership(baseGroup, identity, right.getRole().name());
+			PortfolioRoles role = right.getRole();
+			if(roleSet.contains(role)) {
+				Group baseGroup;
+				if(right.getType() == PortfolioElementType.binder) {
+					baseGroup = binderDao.loadByKey(right.getBinderKey()).getBaseGroup();
+				} else if(right.getType() == PortfolioElementType.section) {
+					baseGroup = binderDao.loadSectionByKey(right.getSectionKey()).getBaseGroup();
+				} else if(right.getType() == PortfolioElementType.page) {
+					baseGroup = pageDao.loadByKey(right.getPageKey()).getBaseGroup();
+				} else {
+					continue;
+				}
+				
+				if(groupDao.hasRole(baseGroup, identity, right.getRole().name())) {
+					groupDao.removeMembership(baseGroup, identity, right.getRole().name());
+				}
 			}
 		}
 	}
@@ -1073,9 +1088,7 @@ public class PortfolioServiceImpl implements PortfolioService {
 		if(StringHelper.containsNonWhitespace(imagePath)) {
 			File bcroot = portfolioFileStorage.getRootDirectory();
 			File file = new File(bcroot, imagePath);
-			if(file.exists()) {
-				file.delete();
-			}
+			FileUtils.deleteFile(file);
 		}
 	}
 
@@ -1430,7 +1443,7 @@ public class PortfolioServiceImpl implements PortfolioService {
 
 		Boolean totalPassed = null;
 		if(totalSectionClosed == assessmentSections.size()) {
-			totalPassed = new Boolean(allPassed);
+			totalPassed = Boolean.valueOf(allPassed);
 		} else {
 			if(assessmentSections.size() == totalSectionPassed) {
 				totalPassed = Boolean.TRUE;
@@ -1443,7 +1456,7 @@ public class PortfolioServiceImpl implements PortfolioService {
 		if("CourseModule".equals(entry.getOlatResource().getResourceableTypeName())) {
 			ICourse course = CourseFactory.loadCourse(entry);
 			CourseNode courseNode = course.getRunStructure().getNode(binder.getSubIdent());
-			ScoreEvaluation scoreEval= new ScoreEvaluation(totalScore.floatValue(), totalPassed, binderStatus, true, null, null, binder.getKey());
+			ScoreEvaluation scoreEval= new ScoreEvaluation(totalScore.floatValue(), totalPassed, binderStatus, true, null, null, null, binder.getKey());
 			UserCourseEnvironment userCourseEnv = AssessmentHelper.createAndInitUserCourseEnvironment(assessedIdentity, course);
 			courseAssessmentService.updateScoreEvaluation(courseNode, scoreEval, userCourseEnv, coachingIdentity, false,
 					Role.coach);
@@ -1497,7 +1510,7 @@ public class PortfolioServiceImpl implements PortfolioService {
 			AssessmentEvaluation eval = courseAssessmentService.getAssessmentEvaluation(pfNode, userCourseEnv);
 			
 			ScoreEvaluation scoreEval = new ScoreEvaluation(eval.getScore(), eval.getPassed(), status, true,
-					null, null, binder.getKey());
+					null, null, null, binder.getKey());
 			courseAssessmentService.updateScoreEvaluation(courseNode, scoreEval, userCourseEnv, coachingIdentity, false,
 					Role.coach);
 		} else {
@@ -1522,8 +1535,7 @@ public class PortfolioServiceImpl implements PortfolioService {
 	}
 
 	private OLATResourceable getOLATResourceableForEvaluationForm(PageBody body) {
-		OLATResourceable ores = OresHelper.createOLATResourceableInstance("portfolio-evaluation", body.getKey());
-		return ores;
+		return OresHelper.createOLATResourceableInstance("portfolio-evaluation", body.getKey());
 	}
 
 	@Override
